+++ src/textAdventure/Key.java
@@ -3,7 +3,7 @@ package textAdventure;
 public class Key extends Item {
   private String keyId;
 
-  public Key(String keyId, String keyName, int weight) {
+  public Key(String keyId, String keyName, long weight) {
     super(weight, keyName);
     this.keyId = keyId;
   }
